docker/requirements.txt
google-cloud-speech==2.25.1
google-cloud-translate>=3.0.0,<4.0.0
tornado==6.3.3